# openapi.yaml
openapi: 3.0.0
info:
  title: Indicator Backend API
  version: 1.0.0
  description: |
    API documentation for the Indicator application backend.

    This API provides endpoints for managing educational institutions' data including:
    - User authentication and management
    - School basic information (alapadatok)
    - Student enrollment data (tanulo_letszam)
    - Educational data (tanugyi_adatok)
    - Competency measurements (kompetencia)
    - Admission statistics (felvettek_szama)
    - Table and data access management

    ## Authentication

    Most endpoints require JWT Bearer token authentication. To authenticate:
    1. Use the `/auth/login` endpoint to obtain a token
    2. Include the token in the Authorization header: `Bearer <token>`

    ## Base URL

    All API endpoints are prefixed with `/api/v1`

  contact:
    name: API Support
    url: https://indicator.pollak.info
    email: support@pollak.info
  license:
    name: Private
    url: https://indicator.pollak.info/license
servers:
  - url: /api/v1
    description: API v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authentication
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
      example:
        error: "Invalid input data"
        message: "The provided data does not meet the required format"
...
    Elegedettseg:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          example: 1
        szakirany_id:
          type: integer
          nullable: true
          description: Field of study identifier
          example: 1
        szakma_id:
          type: integer
          nullable: true
          description: Profession identifier
          example: 1
        alapadatok_id:
          type: integer
          description: School identifier reference
          example: 1
        tanev_kezdete:
          type: integer
          description: School year start year
          example: 2023
        munkaadok_elegedettsege:
          type: number
          format: float
          description: Employer satisfaction rating
          example: 4.2
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          nullable: true
          description: Last update timestamp
    ElegedettsegInput:
      type: object
      required:
        - alapadatok_id
        - tanev_kezdete
        - munkaadok_elegedettsege
      properties:
        szakirany_id:
          type: integer
          nullable: true
          description: Field of study identifier
          example: 1
        szakma_id:
          type: integer
          nullable: true
          description: Profession identifier
          example: 1
        alapadatok_id:
          type: integer
          description: School identifier reference
          example: 1
        tanev_kezdete:
          type: integer
          description: School year start year
          example: 2023
        munkaadok_elegedettsege:
          type: number
          format: float
          description: Employer satisfaction rating
          example: 4.2

    # Versenyek schemas
...
security:
  - bearerAuth: []
...
paths:
...
  /elegedettseg/{tanev}:
    get:
      tags: [Elegedettseg]
      summary: Get satisfaction data by school year
      description: Retrieve employer satisfaction data for a specific school year
      parameters:
        - in: path
          name: tanev
          required: true
          description: The school year end
          schema:
            type: integer
            example: 2024
      responses:
        "200":
          description: Satisfaction data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Elegedettseg"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /elegedettseg/{alapadatokId}/{tanev}:
    get:
      tags: [Elegedettseg]
      summary: Get satisfaction data by school and year
      description: Retrieve employer satisfaction data for a specific school and school year
      parameters:
        - in: path
          name: alapadatokId
          required: true
          description: The school identifier
          schema:
            type: integer
            example: 1
        - in: path
          name: tanev
          required: true
          description: The school year end
          schema:
            type: integer
            example: 2024
      responses:
        "200":
          description: Satisfaction data retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Elegedettseg"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Elegedettseg]
      summary: Delete satisfaction data by school and year
      description: Delete all satisfaction data for a specific school and school year
      parameters:
        - in: path
          name: alapadatokId
          required: true
          description: The school identifier
          schema:
            type: integer
            example: 1
        - in: path
          name: tanev
          required: true
          description: The school year end
          schema:
            type: integer
            example: 2024
      responses:
        "204":
          description: Satisfaction data deleted successfully
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /elegedettseg:
    post:
      tags: [Elegedettseg]
      summary: Create new satisfaction data
      description: Create new employer satisfaction data entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElegedettsegInput"
      responses:
        "201":
          description: Satisfaction data created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Elegedettseg"
        "400":
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /elegedettseg/{id}:
    put:
      tags: [Elegedettseg]
      summary: Update satisfaction data
      description: Update existing employer satisfaction data entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: Satisfaction record ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElegedettsegInput"
      responses:
        "200":
          description: Satisfaction data updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Elegedettseg"
        "400":
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Versenyek endpoints
...